feat: add web crud

+++ src/Controller/Crud/DeleteItemTrait.php
@@ -6,6 +6,7 @@ namespace Siganushka\GenericBundle\Controller\Crud;
 
 use Symfony\Component\HttpFoundation\Response;
 use Symfony\Component\Routing\Attribute\Route;
+use Symfony\Component\Security\Core\Exception\AccessDeniedException;
 
 trait DeleteItemTrait
 {
@@ -15,6 +16,9 @@ trait DeleteItemTrait
     public function deleteItem(string $_id): Response
     {
         $entity = $this->findEntity($_id);
+        if (!$this->isGrantedForOperation(self::OPERATION_DELETE, $entity)) {
+            throw new AccessDeniedException();
+        }
 
         $this->entityManager->remove($entity);
         $this->entityManager->flush();
